requirements.txt
streamlit==1.0.0
streamlit-option-menu==1.0.2
pandas==1.3.0
numpy==1.21.0
matplotlib==3.4.2
scikit-learn==0.24.2
statsmodels==0.12.2
mysql-connector-python==8.0.28
joblib==1.0.1
matplotlib==3.4.2
